Finished inventory can no longer duplicate items
