+++ Assets/Prefabs/Items/ItemTestObjects/Bell.prefab
@@ -12,9 +12,9 @@ GameObject:
   - component: {fileID: 8954813676824300746}
   - component: {fileID: 134737754117262754}
   - component: {fileID: 5797319534187838462}
-  - component: {fileID: 7557835884419092562}
   - component: {fileID: 3847561288180110776}
-  m_Layer: 6
+  - component: {fileID: -5492913113680604131}
+  m_Layer: 7
   m_Name: Bell
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -114,35 +114,6 @@ BoxCollider:
   serializedVersion: 3
   m_Size: {x: 1, y: 1, z: 1}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &7557835884419092562
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8161579611960054099}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 650fc885517557b42b796555f7379dc4, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: Assembly-CSharp::Gameplay.InteractObject
-  _interactEvent:
-    m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 866593153503499546, guid: 42c0db21de0c4e64fbffc4855584c2ad, type: 3}
-        m_TargetAssemblyTypeName: Gameplay.PlayerInventory, Assembly-CSharp
-        m_MethodName: PickupItem
-        m_Mode: 2
-        m_Arguments:
-          m_ObjectArgument: {fileID: 11400000, guid: 1d542ae3372ac6e4fa180ff8d9270385, type: 2}
-          m_ObjectArgumentAssemblyTypeName: InventoryItem, Assembly-CSharp
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
-  _outlineScale: 1.1
-  _outlineMat: {fileID: 2100000, guid: 4d9857703f5614b4db822f7dd9913602, type: 2}
 --- !u!54 &3847561288180110776
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -170,3 +141,21 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &-5492913113680604131
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8161579611960054099}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 23c331980d6b97448b5545cfa6267920, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: Assembly-CSharp::Gameplay.ItemInteractable
+  InteractEvent:
+    m_PersistentCalls:
+      m_Calls: []
+  _outlineScale: 1.1
+  _outlineMat: {fileID: 2100000, guid: 4d9857703f5614b4db822f7dd9913602, type: 2}
+  _item: {fileID: 11400000, guid: 1d542ae3372ac6e4fa180ff8d9270385, type: 2}
