+++ Assets/Prefabs/Items/ItemTestObjects/SpeedSyringe.prefab
@@ -14,7 +14,7 @@ GameObject:
   - component: {fileID: 6004710190513379293}
   - component: {fileID: -7890135874412350714}
   - component: {fileID: 1622018625665191694}
-  m_Layer: 6
+  m_Layer: 7
   m_Name: SpeedSyringe
   m_TagString: Untagged
   m_Icon: {fileID: 0}
