+++ Assets/Scenes/Marinus.unity
@@ -1828,7 +1828,7 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 1883100463497416456, guid: a9aafbc42878d2941909e5e3efa7247d, type: 3}
+    - target: {fileID: 0}
       propertyPath: _interactEvent.m_PersistentCalls.m_Calls.Array.data[0].m_Target
       value: 
       objectReference: {fileID: 1086603889}
@@ -1946,7 +1946,7 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 670581653516023766, guid: fac6a1770667ac946892f2265904918b, type: 3}
+    - target: {fileID: 0}
       propertyPath: _interactEvent.m_PersistentCalls.m_Calls.Array.data[0].m_Target
       value: 
       objectReference: {fileID: 1086603889}
@@ -2104,7 +2104,7 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 7557835884419092562, guid: 7fd5e883c0b6ec1438085b7210672200, type: 3}
+    - target: {fileID: 0}
       propertyPath: _interactEvent.m_PersistentCalls.m_Calls.Array.data[0].m_Target
       value: 
       objectReference: {fileID: 1086603889}
@@ -2169,7 +2169,7 @@ PrefabInstance:
       propertyPath: m_Name
       value: LockPick
       objectReference: {fileID: 0}
-    - target: {fileID: 5409339975775296507, guid: 60ed137dc7a9c3f4eaab24864275afa1, type: 3}
+    - target: {fileID: 0}
       propertyPath: _interactEvent.m_PersistentCalls.m_Calls.Array.data[0].m_Target
       value: 
       objectReference: {fileID: 1086603889}
@@ -2226,7 +2226,7 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 3985151142123665221, guid: 5b978afb698a0c3458b22eed70c5be14, type: 3}
+    - target: {fileID: 0}
       propertyPath: _interactEvent.m_PersistentCalls.m_Calls.Array.data[0].m_Target
       value: 
       objectReference: {fileID: 1086603889}
