+++ Assets/ScriptableObjects/Bell.asset
@@ -9,9 +9,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 6838a444037e5ed4e80916aceaabcfb5, type: 3}
   m_Name: Bell
   m_EditorClassIdentifier: Assembly-CSharp::Bell
   sprite: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
   color: {r: 1, g: 1, b: 1, a: 255}
   HoldObject: {fileID: 8161579611960054099, guid: 7fd5e883c0b6ec1438085b7210672200, type: 3}
+  value: -20
