+++ Assets/ScriptableObjects/Heat Pack.asset
@@ -9,9 +9,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: da638d13933dc3649ba21b2d1228b1bc, type: 3}
   m_Name: Heat Pack
   m_EditorClassIdentifier: Assembly-CSharp::HeatPack
   sprite: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
   color: {r: 1, g: 1, b: 1, a: 255}
   HoldObject: {fileID: 5458958971818635449, guid: 5b978afb698a0c3458b22eed70c5be14, type: 3}
+  value: -20
