+++ Assets/ScriptableObjects/Key.asset
@@ -9,9 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: b8bfbe5fce166d14aae5178817265ad3, type: 3}
   m_Name: Key
   m_EditorClassIdentifier: Assembly-CSharp::Key
   sprite: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
-  color: {r: 0, g: 0.116532326, b: 1, a: 255}
+  color: {r: 1, g: 1, b: 1, a: 255}
   HoldObject: {fileID: 5687983931066264282, guid: a9aafbc42878d2941909e5e3efa7247d, type: 3}
