+++ Assets/ScriptableObjects/LockPick.asset
@@ -9,7 +9,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 94b7b82c058b45d4eaf75562b7009828, type: 3}
   m_Name: Lockpick
   m_EditorClassIdentifier: Assembly-CSharp::Lockpick
   sprite: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
