+++ Assets/ScriptableObjects/Noise Monkey.asset
@@ -9,7 +9,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 2b27676e61103e246b179490da01e53e, type: 3}
   m_Name: Noise Monkey
   m_EditorClassIdentifier: Assembly-CSharp::NoiseMonkey
   sprite: {fileID: 10913, guid: 0000000000000000f000000000000000, type: 0}
