+++ Assets/Scripts/Inventory/Bell.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6838a444037e5ed4e80916aceaabcfb5
