+++ Assets/Scripts/Inventory/Heatpack.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+
+namespace Gameplay
+{
+    [CreateAssetMenu(menuName = "Item/HeatPack")]
+    public class HeatPack : InventoryItem
+    {
+        [SerializeField] private int value = -20;
+        public override bool Use()
+        {
+            PlayerStatusEffects.Instance.AddInstantFrostbite(value);
+            Debug.Log("Heatpack");
+            return true;
+        }
+    }
+}
