+++ Assets/Scripts/Inventory/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e464c34de33d104d98e8924a0c197f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
