+++ Assets/Scripts/Inventory/Key.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b8bfbe5fce166d14aae5178817265ad3
