+++ Assets/Scripts/Inventory/Lockpick.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+namespace Gameplay
+{
+    [CreateAssetMenu(menuName = "Item/Lockpick")]
+    public class Lockpick : InventoryItem
+    {
+        public override bool Use()
+        {
+            Debug.Log("Lockpick");
+            return false;
+            //If door return true
+        }
+    }
+}
