+++ Assets/Scripts/Inventory/Lockpick.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 94b7b82c058b45d4eaf75562b7009828
