+++ Assets/Scripts/Inventory/NoiseMonkey.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+namespace Gameplay
+{
+    [CreateAssetMenu(menuName = "Item/NoiseMonkey")]
+
+    public class NoiseMonkey : InventoryItem
+    {
+      public override bool Use()
+      {
+        Debug.Log("NoiseMonkey");
+        return true;
+      }
+    }
+}
