+++ Assets/Scripts/Inventory/NoiseMonkey.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2b27676e61103e246b179490da01e53e
